feature: Add SupportingContainer
